refactor: fix clippy 1.66 warnings in releaser

+++ utils/releaser/src/apple/mac.rs
@@ -90,7 +90,7 @@ fn upload(gh: &Github) {
     client
         .upload_release_asset(
             &lb_repo(),
-            release.id as u64,
+            release.id,
             "lockbook-macos.app.zip",
             "application/zip",
             file,
